* [html5] fix test case and rebuild some logic
