+++ html5/render/vue/utils/event.js
@@ -35,22 +35,44 @@ function extend (to, ...args) {
  */
 export function createEvent (target, type, props) {
   const event = new Event(type, { bubbles: false })
-  // event.preventDefault()
-  // event.stopPropagation()
 
   extend(event, props)
   //  phantomjs don't support customer event
-  if (window.navigator.userAgent.indexOf('PhantomJS') !== -1) {
+  if (window.navigator.userAgent.toLowerCase().indexOf('phantomjs') !== -1) {
     return event
   }
   try {
     Object.defineProperty(event, 'target', {
       enumerable: true,
-      value: target || null
+      value: target
     })
   }
   catch (err) {
-    return extend({}, event, { target: target || null })
+    return extend({}, event, { target: target })
+  }
+  return event
+}
+
+/**
+ * Create a bubbable Event.
+ * @param {DOMString} type
+ * @param {Object} props
+ */
+export function createBubblesEvent (target, type, props) {
+  const event = new Event(type, { bubbles: true })
+  extend(event, props)
+  //  phantomjs don't support customer event
+  if (window.navigator.userAgent.toLowerCase().indexOf('phantomjs') !== -1) {
+    return event
+  }
+  try {
+    Object.defineProperty(event, 'target', {
+      enumerable: true,
+      value: target
+    })
+  }
+  catch (err) {
+    return extend({}, event, { target: target })
   }
   return event
 }
